+++ packages/create-node-app-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @create-node-app/core
 
+## 0.2.5
+
+### Patch Changes
+
+- Added package managers and custom options
+
 ## 0.2.4
 
 ### Patch Changes